Sandbox restarts crashed processes
